+++ local-config-sample.php
@@ -54,3 +54,8 @@ define( 'WHOOPS_ENABLE', true );
  * If you would like to disable the plugin locally, add the following to your local-config.php.
  */
 define( 'TRIBE_GLOMAR', false );
+
+/*
+ * Set to true to allow force inactive plugins to be enabled.
+ */
+define( 'SQ1_DEBUG', false );